requirements.txt
elevenlabs
elevenlabs[pyaudio]
python-dotenv
requests
pinecone
openai
tqdm
tkinter